fixed wrong type of return variable in ctok()
It should be double, not int

+++ chapters/05/excercises/02.cpp
@@ -8,7 +8,7 @@ them and correct the code
 #include "std_lib_facilities.h"
 double ctok(double c)
 {
-	int k = c - 273.15;
+	double k = c - 273.15;
 	return k;
 }
